+++ src-opam/distro.ml
@@ -312,7 +312,8 @@ type distro =
     | `V36
     | `V37
     | `V38
-    | `V39 ]
+    | `V39
+    | `V40 ]
   | `OracleLinux of [ `V7 | `V8 | `V9 ]
   | `OpenSUSE of
     [ `V42_1
@@ -394,6 +395,7 @@ type t =
     | `V37
     | `V38
     | `V39
+    | `V40
     | `Latest ]
   | `OracleLinux of [ `V7 | `V8 | `V9 | `Latest ]
   | `OpenSUSE of
@@ -636,7 +638,7 @@ let resolve_alias (d : t) : distro =
   | `Alpine `Latest -> `Alpine `V3_19
   | `CentOS `Latest -> `CentOS `V7
   | `Debian `Stable -> `Debian `V12
-  | `Fedora `Latest -> `Fedora `V39
+  | `Fedora `Latest -> `Fedora `V40
   | `OracleLinux `Latest -> `OracleLinux `V9
   | `OpenSUSE `Latest -> `OpenSUSE `V15_5
   | `Ubuntu `Latest -> `Ubuntu `V23_10
@@ -652,7 +654,7 @@ let resolve_alias (d : t) : distro =
     | `Debian (`V7 | `V8 | `V9 | `V10 | `V11 | `V12 | `Testing | `Unstable)
     | `Fedora
         ( `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30
-        | `V31 | `V32 | `V33 | `V34 | `V35 | `V36 | `V37 | `V38 | `V39 )
+        | `V31 | `V32 | `V33 | `V34 | `V35 | `V36 | `V37 | `V38 | `V39 | `V40 )
     | `OracleLinux (`V7 | `V8 | `V9)
     | `OpenSUSE
         ( `V42_1 | `V42_2 | `V42_3 | `V15_0 | `V15_1 | `V15_2 | `V15_3 | `V15_4
@@ -693,7 +695,7 @@ let distro_status (d : t) : status =
         ( `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30
         | `V31 | `V32 | `V33 | `V34 | `V35 | `V36 ) ->
         `Deprecated
-    | `Fedora (`V37 | `V38 | `V39) -> `Active `Tier2
+    | `Fedora (`V37 | `V38 | `V39 | `V40) -> `Active `Tier2
     | `OracleLinux `V7 -> `Deprecated
     | `OracleLinux (`V8 | `V9) -> `Active `Tier3
     | `OpenSUSE
@@ -763,7 +765,7 @@ let distro_arches ov (d : t) =
     when OV.(compare Releases.v4_05_0 ov) = -1 ->
       let base = [ `X86_64; `Aarch64; `Ppc64le; `S390x ] in
       if OV.(compare Releases.v4_11_0 ov) <= 0 then `Riscv64 :: base else base
-  | `Fedora (`V33 | `V34 | `V35 | `V36 | `V37 | `V38 | `V39), ov
+  | `Fedora (`V33 | `V34 | `V35 | `V36 | `V37 | `V38 | `V39 | `V40), ov
     when OV.(compare Releases.v4_08_0 ov) = -1 ->
       [ `X86_64; `Aarch64 ]
   | `OpenSUSE (`V15_4 | `V15_5), ov when OV.(compare Releases.v4_02_0 ov) = -1
@@ -875,6 +877,7 @@ let builtin_ocaml_of_distro (d : t) : string option =
   | `Fedora `V37 -> Some "4.13.1"
   | `Fedora `V38 -> Some "4.14.0"
   | `Fedora `V39 -> Some "5.0.0"
+  | `Fedora `V40 -> Some "5.0.0"
   | `CentOS `V6 -> Some "3.11.2"
   | `CentOS `V7 -> Some "4.01.0"
   | `CentOS `V8 -> Some "4.07.0"
@@ -1022,6 +1025,7 @@ let tag_of_distro (d : t) =
   | `Fedora `V37 -> "fedora-37"
   | `Fedora `V38 -> "fedora-38"
   | `Fedora `V39 -> "fedora-39"
+  | `Fedora `V40 -> "fedora-40"
   | `OracleLinux `V7 -> "oraclelinux-7"
   | `OracleLinux `V8 -> "oraclelinux-8"
   | `OracleLinux `V9 -> "oraclelinux-9"
@@ -1217,6 +1221,7 @@ let human_readable_string_of_distro (d : t) =
     | `Fedora `V37 -> "Fedora 37"
     | `Fedora `V38 -> "Fedora 38"
     | `Fedora `V39 -> "Fedora 39"
+    | `Fedora `V40 -> "Fedora 40"
     | `OracleLinux `V7 -> "OracleLinux 7"
     | `OracleLinux `V8 -> "OracleLinux 8"
     | `OracleLinux `V9 -> "OracleLinux 9"
@@ -1355,6 +1360,7 @@ let bubblewrap_version (t : t) =
   | `Fedora `V37 -> Some (0, 5, 0)
   | `Fedora `V38 -> Some (0, 7, 0)
   | `Fedora `V39 -> Some (0, 8, 0)
+  | `Fedora `V40 -> Some (0, 8, 0)
   | `OracleLinux `V7 -> None
   | `OracleLinux `V8 -> Some (0, 4, 0)
   | `OracleLinux `V9 -> Some (0, 4, 1)
@@ -1494,6 +1500,7 @@ let base_distro_tag ?win10_revision ?(arch = `X86_64) d =
         | `V37 -> "37"
         | `V38 -> "38"
         | `V39 -> "39"
+        | `V40 -> "40"
       in
       ("fedora", tag)
   | `OracleLinux v ->