+++ src-opam/distro.mli
@@ -132,7 +132,8 @@ type distro =
     | `V36
     | `V37
     | `V38
-    | `V39 ]
+    | `V39
+    | `V40 ]
   | `OracleLinux of [ `V7 | `V8 | `V9 ]
   | `OpenSUSE of
     [ `V42_1
@@ -215,6 +216,7 @@ type t =
     | `V37
     | `V38
     | `V39
+    | `V40
     | `Latest ]
   | `OracleLinux of [ `V7 | `V8 | `V9 | `Latest ]
   | `OpenSUSE of